Added multi-threading to core.
Introduced protocol KirinServiceWithUI to make sure we run it on the correct thread.
Fixed up Databases to use a serial queue. When it comes to actually hitting the database, then use either a global dispatch queue (for read only) or serial queue for writing. (SQLite you can only write from one thread at a time, but read is thread safe).