Update the android library project while installing
While installing it now updates the android library project to add files
which ant uses to build from the command line. These usually include the
path which points to where the android SDK is on the machine that kirin
was just installed on. This simplifies the process of using kirin from
the command line by removing a step that you'd usually have to do
manually.