Added mandatory, defaults and allowable.
This also removes the "dependencies" concept of in the validation block of the idl.

The BridgeUtils object contains the schema and is put on the prototype of any generated types. The existence of kirin_buildUtils is a sign that the object can be validated, and if necessary, stored for calling back later on.